\name{collapse}
\alias{collapse}
\title{Alias for frequency.vector function}
\description{This is an alias for the function \code{\link{frequencyvector}}.}
\usage{collapse(D,pos=FALSE)}
\arguments{
  \item{D}{Population vector}
  \item{pos}{Optional: position of categories}
}
\details{This is an alias for the function \code{\link{frequencyvector}}.}
\value{A frequency vector.}
\author{Didier Ruedin}
