\name{GTD2}
\alias{GTD2}
\title{Two-partiness D2}
\description{Calculates the measure of two-partiness D2}
\usage{GTD2(x)}
\arguments{
  \item{x}{Population vector}
}
\details{This function calculates two-partiness (two-partyness) D2, as introduced by Gaines and Taagepera. The argument is a vector with the seat shares of the parties. See \code{\link{GTT}} for the other measure introduced by Gaines and Taaepera.
}
\value{Two-partiness D2}
\author{Didier Ruedin}
\references{Gaines, Brian J., and Rein Taagepera. 2013. How to Operationalize Two-Partyness. Journal of Elections, Public Opinion & Parties. doi:10.1080/17457289.2013.770398.
}
\seealso{\code{\link{GTT}}, \code{\link{neff}}}
