\name{GTT}
\alias{GTT}
\title{Two-partiness T}
\description{Calculates the measure of two-partiness T}
\usage{GTT(x)}
\arguments{
  \item{x}{Population vector}
}
\details{This function calculates two-partiness (two-partyness) T, as introduced by Gaines and Taagepera. The argument is a vector with the seat shares of the parties. See \code{\link{GTD2}} for the other measure introduced by Gaines and Taaepera.
}
\value{Two-partiness T}
\author{Didier Ruedin}
\references{Gaines, Brian J., and Rein Taagepera. 2013. How to Operationalize Two-Partyness. Journal of Elections, Public Opinion & Parties. doi:10.1080/17457289.2013.770398.
}
\seealso{\code{\link{GTD2}}, \code{\link{neff}}}
