\name{polrep-package}
\alias{polrep-package}
\alias{polrep}
\docType{package}
\title{Calculate political representation scores.}
\description{Implementation of different measures of political representation, covering both descriptive representation (unordered groups) and substantive representation (issue positions).}
\details{This package provides different measures of political representation, including potential measures not actually commonly used. It provides individual representation scores (\code{\link{irs}}), the Rose index (\code{\link{Rose}}), and relative representaton scores (\code{\link{relrep}}) among many others, including potential measures of political representation not used in the literature.

Some of the (potential) measures of political representation assume that the largest group in society (i.e. the majority group) is in the first position of the vector. Putting the largest group into position one of the vector is no disadvantage for any of the measures.
}
\author{
Didier Ruedin

Maintainer: Didier Ruedin <didier.ruedin@wolfson.oxon.org>
}
\keyword{ package }
\seealso{\code{\link{irs}}, \code{\link{irs.cat}}, \code{\link{Rose}}, \code{\link{relrep}}, \code{\link{binrep}}, \code{\link{Farina}}, \code{\link{Gallagher}}, \code{\link{Rae}}, \code{\link{StLague}}}

