\name{score.A}
\alias{score.A}
\title{Representation score A}
\description{Calculates the (potential) representation score A.}
\usage{score.A(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum((Z-R)^2).}
\value{Representation score}
\author{Didier Ruedin}
