\name{score.B}
\alias{score.B}
\title{Representation score B}
\description{Calculates the (potential) representation score B.}
\usage{score.B(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum(Z/R).}
\value{Representation score}
\author{Didier Ruedin}
