\name{score.D}
\alias{score.D}
\title{Representation score D}
\description{Calculates the (potential) representation score D.}
\usage{score.D(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum(Z^2/R^2).}
\value{Representation score}
\author{Didier Ruedin}
