\name{score.E}
\alias{score.E}
\title{Representation score "E"}
\description{Calculates the (potential) representation score "E".}
\usage{score.E(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum((Z-R)/Z).}
\value{Representation score}
\author{Didier Ruedin}
