\name{score.H}
\alias{score.H}
\title{Representation score "H"}
\description{Calculates the (potential) representation score "H".}
\usage{score.H(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum(R/Z).}
\value{Representation score}
\author{Didier Ruedin}
