\name{score.J}
\alias{score.J}
\title{Representation score J}
\description{Calculates the (potential) representation score J.}
\usage{score.J(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum((R/Z)^2).}
\value{Representation score}
\author{Didier Ruedin}
