\name{score.L}
\alias{score.L}
\title{Representation score "L"}
\description{Calculates the (potential) representation score "L".}
\usage{score.L(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum(R^2/Z^2).}
\value{Representation score}
\author{Didier Ruedin}
