\name{score.O}
\alias{score.O}
\title{Representation score "O"}
\description{Calculates the (potential) representation score "O".}
\usage{score.O(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum(Z^2/R).}
\value{Representation score}
\author{Didier Ruedin}
