\name{scoreC}
\alias{scoreC}
\title{Representation score C}
\description{Calculates the (potential) representation score C.}
\usage{scoreC(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum((Z/R)^2).}
\value{Representation score}
\author{Didier Ruedin}
