\name{scoreI}
\alias{scoreI}
\title{Representation score I}
\description{Calculates the (potential) representation score I.}
\usage{scoreI(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum(R/Z)/the number of groups.}
\value{Representation score}
\author{Didier Ruedin}
