\name{scoreP}
\alias{scoreP}
\title{Representation score P}
\description{Calculates the (potential) representation score P.}
\usage{scoreP(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as: sum(R/Z^2).}
\value{Representation score}
\author{Didier Ruedin}
