\name{scoreT}
\alias{scoreT}
\title{Representation score T}
\description{Calculates the (potential) representation score T.}
\usage{scoreT(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as the Rose index * heteregeneity of the population.}
\value{Representation score}
\author{Didier Ruedin}
\seealso{\code{\link{H}}, \code{\link{Rose}}}
