\name{scoreU}
\alias{scoreU}
\title{Representation score U}
\description{Calculates the (potential) representation score U.}
\usage{scoreU(Z,R)}
\arguments{
  \item{Z}{Population vector}
  \item{R}{Vector for the representatives}
}
\details{The representation score is calculated as the Rose index / heteregeneity of the population.}
\value{Representation score}
\author{Didier Ruedin}
\seealso{\code{\link{H}}, \code{\link{Rose}}}
