\name{sumZR}
\alias{sumZR}
\title{Calculates sum of absolute differences}
\description{This is a helper function to calculate the sum of the absolute differences |Z-R|
}
\usage{sumZR(Z,R)}
\arguments{
\item{Z}{Population vector}
\item{R}{Vector for the representatives}
}
\details{This is a helper function.}
\value{Sum of absolute differences}
\author{Didier Ruedin}
