\name{Zc}
\alias{Zc}
\title{Calculates proportion of minorities}
\description{This is a helper function to calculate the proportion of minorities in the population}
\usage{Zc(Z)}
\arguments{  \item{Z}{Population vector}}
\details{All groups in the population except for the first one are considered minorities.}
\value{Proportion of minorities}
\author{Didier Ruedin}
