\name{collapse}
\alias{collapse}
\title{Alias for frequency.vector function}
\description{This is an alias for the function \code{\link{frequency.vector}}.}
\usage{collapse(D,k=9999919)}
\arguments{
  \item{D}{Population vector}
  \item{k}{Optional: number of categories}
}
\details{This is an alias for the function \code{\link{frequency.vector}}.}
\value{A frequency vector.}
\author{Didier Ruedin}
