\name{H}
\alias{H}
\title{Calculates heterogeneity}
\description{This function calculates heterogeneity from a population vector.}
\usage{H(Z)}
\arguments{ \item{Z}{Population vector}}
\details{This function implements Fearon's (2003) measure of heterogeneity. Specify a frequency vector.}
\value{Level of heterogeneity}
\author{Didier Ruedin}
\references{Fearon, J. (2003) Ethnic and cultural diversity by country. Journal of Economic Growth 8:195-222.}
