\name{polrep-package}
\alias{polrep-package}
\alias{polrep}
\docType{package}
\title{Calculate political representation scores.}
\description{Implementation of different measures of political representation, covering both descriptive representation (unordered groups) and substantive representation (issue positions).}
\details{This package provides different measures of political representation. Currently it only provides individual representation scores (\code{\link{irs}}).
}
\author{
Didier Ruedin

Maintainer: Didier Ruedin <didier.ruedin@wolfson.oxon.org>
}
\references{}
\keyword{ package }
\seealso{\code{\link{irs}}, \code{\link{irs.cat}}}
\examples{}
